Delete confirmation for manual accounting_entries

+++ app/controllers/destroy_confirmations_controller.rb
@@ -2,15 +2,21 @@ class DestroyConfirmationsController < ApplicationController
   def new
     @obj = parent
     authorize! :destroy, @obj
+    redirect_to "/403.html" if AccountingEntry === @obj && !@obj.manual?
     render :layout => !request.xhr?
   end
 
   def create
     @obj = parent
     authorize! :destroy, @obj
+    redirect_to "/403.html" if AccountingEntry === @obj && !@obj.manual?
     if @obj.destroy
       unless request.xhr?
-        redirect_to @obj.class
+        unless AccountingEntry === @obj
+          redirect_to @obj.class
+        else
+          redirect_to accounting_session_manual_accounting_entries_path(@obj.accounting_session)
+        end
       else
         head :ok
       end
@@ -21,7 +27,7 @@ class DestroyConfirmationsController < ApplicationController
 
 private
   def parent
-    [ :flight ].each do |o|
+    [ :flight, :accounting_entry ].each do |o|
       return o.to_s.camelcase.constantize.find(params["#{o}_id"]) if params["#{o}_id"]
     end
   end
